pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "boxmot>=15.0.5",
    "cleanlab>=2.7.1",
    "datasets>=4.2.0",
    "hf-transfer>=0.1.9",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.6",
    "numpy>=1.26.0,<2.0.0",
    "onnx>=1.19.0",
    "onnxruntime-gpu>=1.23.0",
    "onnxslim>=0.1.71",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pycuda>=2025.1.2",
    "rfdetr>=1.3.0",
    "sahi>=0.11.36",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "torch-tensorrt>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
    "ultralytics>=8.3.205",
    "google-generativeai"
]
